--- api-common/src/main/kotlin/com/kidsworld/api/common/exception/CommonException.kt
@@ -1,8 +1,0 @@
-package com.kidsworld.api.common.exception
-
-import org.springframework.http.HttpStatus
-
-abstract class CommonException(message: String) : Exception(message) {
-    abstract val errorCode: Int
-    abstract val httpStatus: HttpStatus
-}
