--- api-common/src/main/kotlin/com/kidsworld/api/common/exception/UnknownException.kt
@@ -1,8 +1,0 @@
-package com.kidsworld.api.common.exception
-
-import org.springframework.http.HttpStatus
-
-class UnknownException(override val errorCode: Int, override val message: String) : CommonException(message) {
-    override val httpStatus: HttpStatus
-        get() = HttpStatus.SERVICE_UNAVAILABLE
-}
