+++ api-common/src/main/kotlin/com/kidsworld/web/api/common/exception/CommonException.kt
@@ -1,0 +1,8 @@
+package com.kidsworld.web.api.common.exception
+
+import org.springframework.http.HttpStatus
+
+abstract class CommonException(message: String) : Exception(message) {
+    abstract val errorCode: Int
+    abstract val httpStatus: HttpStatus
+}
