+++ api-common/src/main/kotlin/com/kidsworld/web/api/common/exception/UnknownException.kt
@@ -1,0 +1,8 @@
+package com.kidsworld.web.api.common.exception
+
+import org.springframework.http.HttpStatus
+
+class UnknownException(override val errorCode: Int, override val message: String) : CommonException(message) {
+    override val httpStatus: HttpStatus
+        get() = HttpStatus.SERVICE_UNAVAILABLE
+}
