--- api/src/main/kotlin/com/kidsworld/api/ApiApplication.kt
@@ -1,13 +1,0 @@
-package com.kidsworld.api
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-import org.springframework.boot.runApplication
-
-@SpringBootApplication(scanBasePackages = ["com.kidsworld"])
-@ConfigurationPropertiesScan
-class ApiApplication
-
-fun main(args: Array<String>) {
-    runApplication<ApiApplication>(*args)
-}
