--- api/src/main/kotlin/com/kidsworld/api/config/ApiConfig.kt
@@ -1,8 +1,0 @@
-package com.kidsworld.api.config
-
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-//@ComponentScan(basePackages = ["com.kidsworld"])
-class ApiConfig
