+++ api/src/main/kotlin/com/kidsworld/web/api/ApiApplication.kt
@@ -1,0 +1,13 @@
+package com.kidsworld.web.api
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+import org.springframework.boot.runApplication
+
+@SpringBootApplication(scanBasePackages = ["com.kidsworld"])
+@ConfigurationPropertiesScan
+class ApiApplication
+
+fun main(args: Array<String>) {
+    runApplication<ApiApplication>(*args)
+}
