+++ api/src/main/kotlin/com/kidsworld/web/api/config/ApiConfig.kt
@@ -1,0 +1,8 @@
+package com.kidsworld.web.api.config
+
+import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+//@ComponentScan(basePackages = ["com.kidsworld"])
+class ApiConfig
