+++ api/src/test/kotlin/com/kidsworld/web/api/DummyTest.kt
@@ -1,0 +1,15 @@
+package com.kidsworld.web.api
+
+import com.kidsworld.web.domain.util.DummyUtil
+import org.assertj.core.api.Assertions
+import org.junit.jupiter.api.Test
+import org.springframework.test.context.TestConstructor
+
+@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
+class DummyTest() {
+
+    @Test
+    fun `common모듈 내 정적매소드 호출 테스트`() {
+        Assertions.assertThat(DummyUtil.sum(1, 1)).isEqualTo(2)
+    }
+}
