--- domain/src/main/kotlin/com/kidsworld/domain/repository/dev2db/User2Repository.kt
@@ -1,8 +1,0 @@
-package com.kidsworld.domain.repository.dev2db
-
-import com.kidsworld.domain.entity.UserEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface User2Repository: JpaRepository<UserEntity, Long> {
-    fun findTopByName(name: String): UserEntity?
-}
