--- domain/src/main/kotlin/com/kidsworld/domain/util/DummyUtil.kt
@@ -1,11 +1,0 @@
-package com.kidsworld.domain.util
-
-class DummyUtil {
-
-    companion object {
-        fun sum(x: Int, y: Int): Int {
-            return x + y
-        }
-    }
-}
-
