+++ domain/src/main/kotlin/com/kidsworld/web/domain/util/DummyUtil.kt
@@ -1,0 +1,11 @@
+package com.kidsworld.web.domain.util
+
+class DummyUtil {
+
+    companion object {
+        fun sum(x: Int, y: Int): Int {
+            return x + y
+        }
+    }
+}
+
